Move sshfs utility functions to 'aim/ext/sshfs/utils'
